Refactor to shrink and specialize modules
